updated: soft delete all validations

+++ src/main/java/com/bookrentalsystem/bks/controller/author/AuthorController.java
@@ -2,8 +2,11 @@ package com.bookrentalsystem.bks.controller.author;
 
 import com.bookrentalsystem.bks.dto.author.AuthorRequest;
 import com.bookrentalsystem.bks.dto.author.AuthorResponse;
+import com.bookrentalsystem.bks.exception.globalException.AuthorCanNotBeDeletedException;
 import com.bookrentalsystem.bks.model.Author;
+import com.bookrentalsystem.bks.model.Book;
 import com.bookrentalsystem.bks.service.AuthorService;
+import com.bookrentalsystem.bks.service.BookService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -15,13 +18,17 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/author")
 @RequiredArgsConstructor
 public class AuthorController {
     private final AuthorService authorService;
+    private final BookService bookService;
 
+    //author table
     @GetMapping("/table")
     public String authorTable(Model model){
       List<AuthorResponse> authors = authorService.allAuthor();
@@ -29,6 +36,7 @@ public class AuthorController {
         return "author/AuthorTable";
     }
 
+    //author form page
     @GetMapping("/form")
     public String authorForm(Model model){
         if(model.getAttribute("author") == null) {
@@ -36,7 +44,7 @@ public class AuthorController {
         }
         return "author/AuthorForm";
     }
-
+//save author in the db
     @PostMapping("/save")
     public String saveAuthor(@Valid @ModelAttribute("author") AuthorRequest authorRequest,
                              BindingResult result,
@@ -70,6 +78,7 @@ public class AuthorController {
       return "/author/UpdateForm";
     }
 
+    //update author controller
     @PostMapping("/update/save")
     public String saveUpdateAuthor(@Valid @ModelAttribute("author") AuthorRequest authorRequest,
                              BindingResult result,
@@ -85,9 +94,28 @@ public class AuthorController {
 
         return "redirect:/author/table";
     }
+
+    //delete author controller
     @RequestMapping("/delete/{id}")
-    public String deleteCategory(@PathVariable Short id){
-        authorService.deleteAuthor(id);
+    public String deleteAuthor(@PathVariable Short id,RedirectAttributes redirectAttributes){
+        List<Book> allBooks = bookService.allBookEntity();
+        List<Author> allAuthors = authorService.allAuthors();
+
+       List<Short> allIds = allBooks.stream()
+               .flatMap(b -> b.getAuthors().stream())
+               .map(Author::getId).collect(Collectors.toList());
+
+       Optional<Short> singleId = allIds.stream().filter(i -> i == id).findFirst();
+
+       //if the id is present in the book then it will throw exception
+       if(singleId.isPresent()){
+           throw new AuthorCanNotBeDeletedException("Author cannot be deleted!!");
+       }
+
+       authorService.deleteAuthor(id);
+
+        String message = "";
+        redirectAttributes.addFlashAttribute("message","Author Deleted Successfully!!");
         return "redirect:/author/table?success";
     }
